Refactor: 폴더구조 변경, 사용하지 않는 import 삭제
