+++ src/components/DashBoard/Widget/Chart/BarChart/type.d.ts
